+++ packages/vite-plugin-web-extension/src/constants.ts
@@ -2,4 +2,3 @@ export const MANIFEST_LOADER_PLUGIN_NAME = `web-extension:manifest`;
 export const LABELED_STEP_PLUGIN_NAME = `web-extension:labeled-step`;
 export const MULTIBUILD_COMPLETE_PLUGIN_NAME = `web-extension:multibuild`;
 export const BUNDLE_TRACKER_PLUGIN_NAME = `web-extension:bundle-tracker`;
-export const DISABLE_DEV_PLUGIN_NAME = `web-extension:disable-dev`;
