+++ packages/vite-plugin-web-extension/src/extension-runner/index.ts
@@ -1,0 +1,2 @@
+export { ExtensionRunner } from "./interface";
+export { createWebExtRunner, WebExtRunnerOptions } from "./web-ext-runner";
