+++ packages/vite-plugin-web-extension/src/extension-runner/interface.ts
@@ -1,0 +1,5 @@
+export interface ExtensionRunner {
+  openBrowser(): Promise<void>;
+  reload(): Promise<void>;
+  exit(): Promise<void>;
+}
