+++ packages/vite-plugin-web-extension/src/logger.ts
