--- packages/vite-plugin-web-extension/src/utils/arrays.ts
@@ -1,3 +1,0 @@
-export function compact<T>(array: Array<T | undefined>): T[] {
-  return array.filter((item) => item != null) as T[];
-}
